requirements.txt
python-mcp>=1.0.0
sqlmodel>=0.0.8
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pytest>=7.0.0
uvicorn>=0.20.0
pydantic>=2.0.0